+++ ft_memcpy.c
@@ -1,18 +1,30 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_memcpy.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: hatalhao <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/11/03 17:29:38 by hatalhao          #+#    #+#             */
+/*   Updated: 2023/11/03 17:31:34 by hatalhao         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
-void    *ft_memcpy(void *dest, const void *src, unsigned int n)
+void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-    unsigned int    i;
-    char    *str1;
-    char    *str2;
+	unsigned int	i;
+	char			*str1;
+	char			*str2;
 
-    str1 = (char*) src;
-    str2 = (char*) dest;
-    i = 0;
-    while (str1[i] && i < n)
-    {
-        str2[i] = str1[i];
-        i++;
-    }
-    return (str2);
+	str1 = (char *) src;
+	str2 = (char *) dest;
+	i = 0;
+	while (str1[i] && i < n)
+	{
+		str2[i] = str1[i];
+		i++;
+	}
+	return (str2);
 }
