+++ ft_memset.c
@@ -1,18 +1,30 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_memset.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: hatalhao <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/11/03 17:34:33 by hatalhao          #+#    #+#             */
+/*   Updated: 2023/11/03 17:35:33 by hatalhao         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
-void    *ft_memset(void *s, int c, unsigned int n)
+void	*ft_memset(void *s, int c, size_t n)
 {
-    unsigned char *str;
-    unsigned char a;
-    unsigned int i;
+	unsigned char	*str;
+	unsigned char	a;
+	unsigned int	i;
 
-    str = s;
-    a = c;
-    i = 0;
-    while (i < n)
-    {
-        str[i] = a;
-        i++;
-    }
-    return (str);
+	str = s;
+	a = c;
+	i = 0;
+	while (i < n)
+	{
+		str[i] = a;
+		i++;
+	}
+	return (str);
 }
