+++ ft_strlcpy.c
@@ -1,20 +1,32 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strlcpy.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: hatalhao <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/11/03 17:55:01 by hatalhao          #+#    #+#             */
+/*   Updated: 2023/11/03 17:56:49 by hatalhao         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
-unsigned int    ft_strlcpy(char *dest, const char *src, unsigned int size)
+size_t	ft_strlcpy(char *dest, const char *src, size_t size)
 {
-    unsigned int    i;
-    unsigned int    len;
-    char    *str;
+	unsigned int	i;
+	unsigned int	len;
+	char			*str;
 
-    str = (char *) src;
-    i = 0;
-    len = ft_strlen(str);
-    if (size == 0)
-        return (len);
-    while (str[i] && i < size)
-    {
-        dest[i] = str[i];
-        i++;
-    }
-    return (len);
+	str = (char *) src;
+	i = 0;
+	len = ft_strlen(str);
+	if (size == 0)
+		return (len);
+	while (str[i] && i < size)
+	{
+		dest[i] = str[i];
+		i++;
+	}
+	return (len);
 }
